requirements.txt
Pillow
face_recognition